requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langgraph
langchain-openai
langchain-ollama
qdrant-client
requests
beautifulsoup4
html2text
lxml
tiktoken
langchain_deepseek
langchain-community
langchain-core
langchain-text-splitters
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
# Premium format generation packages
markdown
weasyprint
python-docx
jinja2
# Additional dependencies for enhanced functionality
Pillow
reportlab
cssselect2
tinycss2
cffi
pycparser